Pass arguments to configure script and use cc instead of ld
When building spdk with --enable-debug there was a problem with undefined
fstat64 symbol. Deeper investigation revealed that libc.so does not define
it and we need gcc to add a pixie dust to redirect from fstat64 to a
similar function defined in libc before the object files are handed over
to a linker.